requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
cycler==0.10.0
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
findiff==0.8.5
hyperlink==20.0.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.1.0
incremental==17.5.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
multitasking==0.0.9
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nose==1.3.7
notebook==6.1.5
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-binance==0.7.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyWavelets==1.1.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.0
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
service-identity==18.1.0
simplification==0.5.7
six==1.15.0
sympy==1.6.2
terminado==0.9.1
testpath==0.4.4
tifffile==2020.11.26
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
